pyproject.toml
[tool.poetry]
name = "podcast-links-manager"
version = "0.1.0"
description = "Telegram bot for retrieving the links of a new episode in all configured platforms"
authors = ["Ale Sanchez Medina <hi@alesanchez.es>"]
license = "The Unlicense"

[tool.poetry.dependencies]
python = "^3.10.7"
python-telegram-bot = "^13.11"
python-dotenv = "^0.21.0"
spotipy = "^2.20.0"
pymongo = "^4.2.0"
pycountry = "^22.3.5"

[tool.poetry.dev-dependencies]
pylint = "^2.12.2"
black = "^22.1.0"
isort = "^5.10.1"
flake8 = "^4.0.1"
bandit = "^1.7.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
